Worked on the LXC quick start
Side note: LXC is a complete pain. I could only get it to work reliably on the newest Ubuntu, and even then there were still a couple of times that I had to reboot the VM because something got completely screwed up or there was a process that wouldn't die or something. So that's why LXD is its own section; it's definitely easier to use and people actually run it in production.